Try waiting until all analyses complete.
